# Plan/client/swagger-docs-3.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NoteEarly API Documentation (Zod-Generated)",
      "version": "1.0.0",
      "description": "API documentation for the NoteEarly platform (schemas generated from Zod)."
    },
    "servers": [
      {
        "url": "http://localhost:4000/api/v1",
        "description": "Development server (V1)"
      }
    ],
    "components": {
...
      "schemas": {
...
        "StartModuleInput": {
          "type": "object",
          "properties": {
            "moduleId": {
              "type": "string",
              "format": "uuid"
            }
          },
          "required": [
            "moduleId"
          ]
        },
        "SubmitSummaryInput": {
          "type": "object",
          "properties": {
            "moduleId": {
              "type": "string",
              "format": "uuid"
            },
            "paragraphIndex": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true
            },
            "paragraphSummary": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1000
            },
            "cumulativeSummary": {
              "type": "string",
              "minLength": 1,
              "maxLength": 10000
            }
          },
          "required": [
            "moduleId",
            "paragraphIndex",
            "paragraphSummary",
            "cumulativeSummary"
          ]
        },
...
        "AdminUpdateProgressInputSchema": {
          "type": "object",
          "properties": {
            "score": {
              "type": "integer",
              "nullable": true,
              "minimum": 0,
              "maximum": 100
            },
            "teacherFeedback": {
              "type": "string",
              "nullable": true,
              "maxLength": 2000
            },
            "completed": {
              "type": "boolean"
            }
          }
        },
        "StudentProgressSchema": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "studentId": {
              "type": "string",
              "format": "uuid"
            },
            "moduleId": {
              "type": "string",
              "format": "uuid"
            },
            "completed": {
              "type": "boolean"
            },
            "score": {
              "type": "integer",
              "nullable": true
            },
            "highestParagraphIndexReached": {
              "type": "integer",
              "nullable": true
            },
            "finalSummary": {
              "type": "string",
              "nullable": true
            },
            "startedAt": {
              "type": "string",
              "nullable": true
            },
            "completedAt": {
              "type": "string",
              "nullable": true
            },
            "timeSpentMinutes": {
              "type": "integer",
              "nullable": true
            },
            "createdAt": {
              "type": "string"
            },
            "updatedAt": {
              "type": "string"
            },
            "teacherFeedback": {
              "type": "string",
              "nullable": true
            },
            "teacherFeedbackAt": {
              "type": "string",
              "nullable": true
            }
          },
          "required": [
            "id",
            "studentId",
            "moduleId",
            "completed",
            "score",
            "highestParagraphIndexReached",
            "finalSummary",
            "startedAt",
            "completedAt",
            "timeSpentMinutes",
            "createdAt",
            "updatedAt",
            "teacherFeedback",
            "teacherFeedbackAt"
          ]
        },
...
        "StudentProgressDetailsDTOSchema": {
          "type": "object",
          "properties": {
            "progress": {
              "type": "object",
              "nullable": true,
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "studentId": {
                  "type": "string",
                  "format": "uuid"
                },
                "moduleId": {
                  "type": "string",
                  "format": "uuid"
                },
                "completed": {
                  "type": "boolean"
                },
                "score": {
                  "type": "integer",
                  "nullable": true
                },
                "highestParagraphIndexReached": {
                  "type": "integer",
                  "nullable": true
                },
                "finalSummary": {
                  "type": "string",
                  "nullable": true
                },
                "startedAt": {
                  "type": "string",
                  "nullable": true
                },
                "completedAt": {
                  "type": "string",
                  "nullable": true
                },
                "timeSpentMinutes": {
                  "type": "integer",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "teacherFeedback": {
                  "type": "string",
                  "nullable": true
                },
                "teacherFeedbackAt": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "studentId",
                "moduleId",
                "completed",
                "score",
                "highestParagraphIndexReached",
                "finalSummary",
                "startedAt",
                "completedAt",
                "timeSpentMinutes",
                "createdAt",
                "updatedAt",
                "teacherFeedback",
                "teacherFeedbackAt"
              ]
            },
            "submissions": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "studentProgressId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "paragraphIndex": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "paragraphSummary": {
                    "type": "string"
                  },
                  "cumulativeSummary": {
                    "type": "string"
                  },
                  "submittedAt": {
                    "type": "string"
                  },
                  "createdAt": {
                    "type": "string"
                  },
                  "updatedAt": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "studentProgressId",
                  "paragraphIndex",
                  "paragraphSummary",
                  "cumulativeSummary",
                  "submittedAt",
                  "createdAt",
                  "updatedAt"
                ]
              }
            }
          },
          "required": [
            "progress",
            "submissions"
          ]
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "Enter JWT token **_only_**"
        }
      },
      "parameters": {
        "ModuleIdParam": {
          "in": "path",
          "name": "moduleId",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "The UUID of the module."
        },
        "ProgressIdParam": {
          "in": "path",
          "name": "progressId",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "The UUID of the student progress record."
        },
...
        "StudentIdParam": {
          "in": "path",
          "name": "studentId",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "The UUID of the student."
        }
      }
    },
    "paths": {
...
      "/api/v1/progress/start": {
        "post": {
          "tags": [
            "Progress - Student"
          ],
          "summary": "Start progress on a module (Student)",
          "description": "Creates a progress record for the authenticated student on a specific module if one doesn't exist. Idempotent.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartModuleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Progress tracking started or retrieved successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "progress": {
                        "$ref": "#/components/schemas/StudentProgress"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error."
            },
            "401": {
              "description": "Authentication required (Student context)."
            },
            "404": {
              "description": "Module or Student not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/submit-summary": {
        "post": {
          "tags": [
            "Progress - Student"
          ],
          "summary": "Submit a paragraph summary (Student)",
          "description": "Submits an individual and cumulative summary for a specific paragraph within a module the student has started.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitSummaryInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Summary submitted successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "submissionId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "progressStatus": {
                        "type": "object",
                        "properties": {
                          "completed": {
                            "type": "boolean"
                          },
                          "highestParagraphIndexReached": {
                            "type": "integer",
                            "nullable": true
                          },
                          "finalSummary": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error, or Module already completed, or paragraph index mismatch."
            },
            "401": {
              "description": "Authentication required (Student context)."
            },
            "404": {
              "description": "Module or Progress record not found for this student."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/details/{moduleId}": {
        "get": {
          "tags": [
            "Progress - Student"
          ],
          "summary": "Get my detailed progress for a module (Student)",
          "description": "Retrieves the overall progress record and all paragraph submissions for the authenticated student on a specific module.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/ModuleIdParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Detailed progress information for the student.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/StudentProgressDetailsDTOSchema"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid Module ID format."
            },
            "401": {
              "description": "Authentication required (Student context)."
            },
            "404": {
              "description": "Progress record not found for this student/module."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/my-progress": {
        "get": {
          "tags": [
            "Progress - Student"
          ],
          "summary": "Get all my progress records (Student)",
          "description": "Retrieves a list of all progress records (across different modules) for the currently authenticated student.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "A list of the student's progress records.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/StudentProgress"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Authentication required (Student context)."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/admin/update/{progressId}": {
        "patch": {
          "tags": [
            "Progress - Admin"
          ],
          "summary": "Update progress record (Admin/SuperAdmin)",
          "description": "Allows an Admin/SuperAdmin to update specific fields (e.g., feedback, score, completion status) on a student's progress record they manage.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/ProgressIdParam"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateProgressInputSchema"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Progress record updated successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/StudentProgress"
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error or Invalid ID."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Forbidden (Admin/SuperAdmin access required, or does not manage the student associated with this progress)."
            },
            "404": {
              "description": "Progress record not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/admin/module/{moduleId}": {
        "get": {
          "tags": [
            "Progress - Admin"
          ],
          "summary": "Get all progress for a module (Admin/SuperAdmin)",
          "description": "Retrieves all student progress records associated with a specific module, accessible only to Admins/SuperAdmins.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/ModuleIdParam"
            }
          ],
          "responses": {
            "200": {
              "description": "A list of all progress records for the specified module.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/StudentProgressSchema"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid Module ID format."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Admin/SuperAdmin access required."
            },
            "404": {
              "description": "Module not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/progress/admin/student/{studentId}": {
        "get": {
          "tags": [
            "Progress - Admin"
          ],
          "summary": "Get all progress records for a specific student (Admin)",
          "description": "Retrieves a list of all `StudentProgressSchema` records for a specific student, accessible only by an authorized Admin/SuperAdmin who manages that student.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/StudentIdParam"
            }
          ],
          "responses": {
            "200": {
              "description": "A list of the student's progress records.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/StudentProgressSchema"
                    }
                  }
                }
              }
            },
            "400": { "$ref": "#/components/responses/BadRequest" },
            "401": { "$ref": "#/components/responses/Unauthorized" },
            "403": { "$ref": "#/components/responses/Forbidden" },
            "404": { "$ref": "#/components/responses/NotFound" },
            "500": { "$ref": "#/components/responses/InternalServerError" }
          }
        }
      },
      "/api/v1/progress/admin/student/{studentId}/module/{moduleId}": {
        "get": {
          "tags": [
            "Progress - Admin"
          ],
          "summary": "Get detailed progress for a specific student/module (Admin)",
          "description": "Retrieves the detailed progress (`StudentProgressDetailsDTOSchema`), including all paragraph submissions, for a specific student on a specific module. Accessible only by an authorized Admin/SuperAdmin who manages that student.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "$ref": "#/components/parameters/StudentIdParam"
            },
            {
              "$ref": "#/components/parameters/ModuleIdParam"
            }
          ],
          "responses": {
            "200": {
              "description": "Detailed progress information for the student on the module.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/StudentProgressDetailsDTOSchema"
                  }
                }
              }
            },
            "400": { "$ref": "#/components/responses/BadRequest" },
            "401": { "$ref": "#/components/responses/Unauthorized" },
            "403": { "$ref": "#/components/responses/Forbidden" },
            "404": { "$ref": "#/components/responses/NotFound" },
            "500": { "$ref": "#/components/responses/InternalServerError" }
          }
        }
      },
...
    },
...
  }